cut file last

+++ 02_programming_fundamentals/07_file_management/01_basic_fs/04_cut_file.js
@@ -10,44 +10,53 @@ const { COPYFILE_EXCL } = fs.constants;
 
 function cutPaste(sourceFilename, targetFilename){
 
+
+  if(!fs.existsSync(sourceFilename)){
+    console.warn("File does not exist");
+    return;
+  }
+
   fs.stat(sourceFilename, (err,stats)=>{
     if(err){
-      console.log("error");
-    }else{
-      console.log(stats.isDirectory(sourceFilename));
-      if(stats.isDirectory(sourceFilename)){
-        console.log("We don't want to cut a repertory");
-        return;
-      }
-    }
-  });
-
-
-  fs.copyFile(sourceFilename, targetFilename, COPYFILE_EXCL,(err) => {
-    if (err){
-      if(err.code==="EEXIST"){
-        console.log (`File ${targetFilename} already exists. We don't want to override it`);
+      console.warn("error");
+      return;
+    }else if(stats.isDirectory(sourceFilename)){
+        console.warn("We don't want to cut a repertory");
         return;
       }else{
-        console.log (err);
-        return;
+        fs.copyFile(sourceFilename, targetFilename, COPYFILE_EXCL,(err) => {
+          if (err){
+            if(err.code==="EEXIST"){
+              console.warn (`File ${targetFilename} already exists. We don't want to override it`);
+              return;
+            }else{
+              console.warn (err);
+              return;
+            }
+          }
+
+          else{
+            //console.log(`${sourceFilename} was copied to ${targetFilename}`);
+
+            fs.unlink("./"+sourceFilename, (err) => {
+              if (err) throw err;
+              console.log(`${sourceFilename} was cut to ${targetFilename}`);
+            });
+
+            return;
+          }
+        });
+
+
       }
     }
-
-    else{
-      //console.log(`${sourceFilename} was copied to ${targetFilename}`);
-
-      fs.unlink("./"+sourceFilename, (err) => {
-        if (err) throw err;
-        console.log(`${sourceFilename} was cut to ${targetFilename}`);
-      });
-
-      return;
-    }
   });
+
+  //If it's a file..
+
 
 } // END OF copyPaste function
 
-cutPaste ("testText2.txt", "testText6.txt");
+//cutPaste ("testText2.txt", "testText6.txt");
 
 module.exports = cutPaste;
